split out message.go
